% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_append_session_failed.R
\name{append_failed_session_api}
\alias{append_failed_session_api}
\title{Append a failed session to the DB with user info}
\usage{
append_failed_session_api(session_time_failed = Sys.time(), user_info = NA)
}
\arguments{
\item{user_info}{}
}
\description{
Append a failed session to the DB with user info
}
