% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_append_singpause_survey_completion.R
\name{append_singpause_survey_completion}
\alias{append_singpause_survey_completion}
\title{Append singpause survey completion}
\usage{
append_singpause_survey_completion(
  user_id,
  singpause_id,
  psychTestR_id,
  type = c("pretest", "posttest", "start_pretest")
)
}
\arguments{
\item{type}{}
}
\description{
Append singpause survey completion
}
