% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_append_singpause_survey_completion.R
\name{append_singpause_survey_completion_api}
\alias{append_singpause_survey_completion_api}
\title{Append singpause survey completion API}
\usage{
append_singpause_survey_completion_api(
  user_id,
  singpause_id,
  psychTestR_id,
  type = c("pretest", "start_pretest", "posttest")
)
}
\arguments{
\item{type}{}
}
\description{
Append singpause survey completion API
}
