% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{compile_item_trials}
\alias{compile_item_trials}
\title{Compile item trials}
\usage{
compile_item_trials(
  db_con = NULL,
  current_test_id = NULL,
  session_ids_filter = NULL,
  user_ids_filter,
  join_item_banks_on = FALSE,
  filter_item_banks = NULL,
  add_trial_scores = FALSE,
  score_to_use = "opti3",
  trial_filter_fun = NULL,
  grepl_item_id_filter = NULL
)
}
\arguments{
\item{grepl_item_id_filter}{}
}
\description{
Compile item trials
}
