% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_append_session_scores.R
\name{compute_session_scores_and_end_session_api}
\alias{compute_session_scores_and_end_session_api}
\title{Compute session scores and end session API}
\usage{
compute_session_scores_and_end_session_api(
  test_id = NA,
  session_id,
  user_id,
  psychTestR_session_id = NA,
  session_complete = c("0", "1")
)
}
\arguments{
\item{session_complete}{}
}
\description{
Compute session scores and end session API
}
