% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{connect_to_db_state}
\alias{connect_to_db_state}
\title{Connect to the musicassessrdb and store the connection in the state object.}
\usage{
connect_to_db_state(state, ...)
}
\arguments{
\item{...}{}
}
\description{
Connect to the musicassessrdb and store the connection in the state object.
}
