% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_append_demographics.R
\name{db_append_demographics}
\alias{db_append_demographics}
\title{Append a condition to the conditions table}
\usage{
db_append_demographics(db_con, user_id, demographics_df)
}
\arguments{
\item{db_con}{The DB connection.}

\item{user_id}{The user ID.}

\item{demographics_df}{A 1 x 86 dataframe of the demographic information.}
}
\description{
Append a condition to the conditions table
}
