% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_add_trial_and_compute_trial_scores.R
\name{db_append_melodic_production}
\alias{db_append_melodic_production}
\title{Store melodic production data in melodic_production table}
\usage{
db_append_melodic_production(
  db_con,
  trial_id,
  pyin_res,
  correct_boolean,
  correct_boolean_octaves_allowed
)
}
\arguments{
\item{correct_boolean_octaves_allowed}{}
}
\description{
Store melodic production data in melodic_production table
}
