% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_append_session_scores.R
\name{db_append_scores_session}
\alias{db_append_scores_session}
\title{Store session-level scores in scores_session}
\usage{
db_append_scores_session(db_con, session_scores_df)
}
\arguments{
\item{session_scores_df}{Should contain session_id, measure and score columns.}
}
\description{
Store session-level scores in scores_session
}
