% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_add_trial_and_compute_trial_scores.R,
%   R/lambda_midi_add_trial_and_compute_trial_scores.R
\name{db_append_scores_trial}
\alias{db_append_scores_trial}
\title{Append trial-level scores to the scores_trial table}
\usage{
db_append_scores_trial(db_con, trial_id, measure, score)

db_append_scores_trial(db_con, trial_id, measure, score)
}
\arguments{
\item{score}{}
}
\description{
Append trial-level scores to the scores_trial table

Append trial-level scores to the scores_trial table
}
