% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_append_session_tokens.R
\name{db_append_session_tokens}
\alias{db_append_session_tokens}
\title{Append a session token to the session_tokens table}
\usage{
db_append_session_tokens(user_id, expiration_in_seconds = 600)
}
\arguments{
\item{expiration_in_seconds}{}
}
\description{
Append a session token to the session_tokens table
}
