% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_append_tests.R
\name{db_append_tests}
\alias{db_append_tests}
\title{Append an experiment to the tests table}
\usage{
db_append_tests(db_con, test_name, test_description, active = TRUE)
}
\arguments{
\item{db_con}{The DB connection.}

\item{test_name}{The name of the test.}

\item{test_description}{A description of the test.}

\item{active}{}
}
\description{
Append an experiment to the tests table
}
