% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{db_append_to_table}
\alias{db_append_to_table}
\title{Append a data frame to a table in the database}
\usage{
db_append_to_table(db_con, table, data, primary_key_col = NULL)
}
\arguments{
\item{db_con}{A database connection environment}

\item{table}{Name of table to which to append in database}

\item{data}{dataframe to be appended}
}
\value{
Returns number of rows affected
}
\description{
Append a data frame to a table in the database
}
