% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_append_user_instrument_info.R
\name{db_append_user_instrument_info}
\alias{db_append_user_instrument_info}
\title{Append user instrument information the the user_instrument_info table}
\usage{
db_append_user_instrument_info(
  db_con,
  user_id,
  instrument_id,
  bottom_range,
  top_range
)
}
\arguments{
\item{top_range}{}
}
\description{
Append user instrument information the the user_instrument_info table
}
