% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_append_users.R
\name{db_append_users}
\alias{db_append_users}
\title{Append a user to the users table}
\usage{
db_append_users(
  db_con,
  username,
  password,
  enabled = TRUE,
  created_at = Sys.time()
)
}
\arguments{
\item{created_at}{}
}
\description{
Append a user to the users table
}
