% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{elt_add_final_session_info_to_db}
\alias{elt_add_final_session_info_to_db}
\title{Add final session info to postgres db}
\usage{
elt_add_final_session_info_to_db(asynchronous_api_mode)
}
\arguments{
\item{asynchronous_api_mode}{}
}
\description{
Add final session info to postgres db
}
