% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{elt_disconnect_from_db}
\alias{elt_disconnect_from_db}
\title{Disconnect from the musicassessrdb, as stored in the state object}
\usage{
elt_disconnect_from_db()
}
\description{
Disconnect from the musicassessrdb, as stored in the state object
}
