% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{get_available}
\alias{get_available}
\title{Get tests or experiments available to a user}
\usage{
get_available(db_con, user_id, type = c("tests", "experiments"))
}
\arguments{
\item{db_con}{Db connection}

\item{user_id}{User ID}
}
\value{
Data.frame with id and name
}
\description{
Get tests or experiments available to a user
}
