% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_select_items.R
\name{get_job_status_api}
\alias{get_job_status_api}
\title{Get job status API for select items lambda}
\usage{
get_job_status_api(job_id = NULL, filename = NULL)
}
\arguments{
\item{filename}{}
}
\description{
Get job status API for select items lambda
}
