% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jobs_api.R
\name{get_jobs_api}
\alias{get_jobs_api}
\title{Compute session scores and end session API}
\usage{
get_jobs_api(filename)
}
\arguments{
\item{filename}{}
}
\description{
Compute session scores and end session API
}
