% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{get_range_from_user_id}
\alias{get_range_from_user_id}
\title{Get a user's range based on their user ID}
\usage{
get_range_from_user_id(db_con, user_id, instrument_id = 1L)
}
\arguments{
\item{user_id}{}
}
\description{
Get a user's range based on their user ID
}
