% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permissions.R
\name{give_user_permissions}
\alias{give_user_permissions}
\title{Give a specific user permissions to a test or experiment.}
\usage{
give_user_permissions(
  db_con,
  user_id,
  id,
  type = c("tests", "experiments"),
  action = c("Allow", "Deny")
)
}
\arguments{
\item{id}{The test or experiment ID.}

\item{action}{}
}
\description{
Give a specific user permissions to a test or experiment.
}
