% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{instrument_name_to_id}
\alias{instrument_name_to_id}
\title{Get instrument ID from name}
\usage{
instrument_name_to_id(inst_name)
}
\arguments{
\item{inst_name}{}
}
\description{
Get instrument ID from name
}
