% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/lambda_midi_add_trial_and_compute_trial_scores.R
\name{midi_add_trial_and_compute_trial_scores_api}
\alias{midi_add_trial_and_compute_trial_scores_api}
\title{Add MIDI trial via API}
\usage{
midi_add_trial_and_compute_trial_scores_api(
  stimuli,
  stimuli_durations,
  test_id,
  item_id,
  user_id,
  instrument,
  trial_time_started,
  trial_time_completed,
  score_to_use = "opti3",
  display_modality = "auditory",
  phase,
  rhythmic,
  session_id,
  review_items_id = NA,
  new_items_id = NA,
  dur,
  onset,
  note,
  attempt,
  additional,
  melody_block_paradigm = NA,
  page_label = "",
  module = "NA"
)
}
\arguments{
\item{module}{}
}
\description{
Add MIDI trial via API
}
