% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{musicassessr_con}
\alias{musicassessr_con}
\title{Connect to musicassessr DB}
\usage{
musicassessr_con(local = FALSE, db_name = Sys.getenv("MUSICASSESSR_DB_NAME"))
}
\arguments{
\item{db_name}{}
}
\description{
Connect to musicassessr DB
}
