% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{musicassessr_shiny_init}
\alias{musicassessr_shiny_init}
\title{Initiate a connection to the musicassessr db}
\usage{
musicassessr_shiny_init(on_stop = TRUE, connect_to_db = TRUE)
}
\arguments{
\item{connect_to_db}{}
}
\description{
Initiate a connection to the musicassessr db
}
