% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{musicassessr_shiny_on_stop}
\alias{musicassessr_shiny_on_stop}
\title{Close up connections to the musicassessr db}
\usage{
musicassessr_shiny_on_stop()
}
\description{
Close up connections to the musicassessr db
}
