% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_sample_from_item_bank.R
\name{sample_from_item_bank_api}
\alias{sample_from_item_bank_api}
\title{Sample from an item bank via the API}
\usage{
sample_from_item_bank_api(
  item_bank_name,
  num_items,
  span = NULL,
  melody_length = NULL
)
}
\arguments{
\item{melody_length}{}
}
\description{
Sample from an item bank via the API
}
