% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_sample_from_item_bank.R
\name{sample_from_item_bank_elts}
\alias{sample_from_item_bank_elts}
\title{Sample from item bank via API}
\usage{
sample_from_item_bank_elts(
  item_bank_name = "WJD_ngram",
  num_items,
  melody_length
)
}
\arguments{
\item{melody_length}{}
}
\description{
Sample from item bank via API
}
