% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_sample_melody_in_key.R
\name{sample_melody_in_key_elts}
\alias{sample_melody_in_key_elts}
\title{Sample from item bank via API}
\usage{
sample_melody_in_key_elts(
  item_bank_name = "WJD_ngram",
  num_items,
  key_difficulty,
  melody_length
)
}
\arguments{
\item{melody_length}{}
}
\description{
Sample from item bank via API
}
