% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_user_entry_into_test.R
\name{validate_user_entry_into_test}
\alias{validate_user_entry_into_test}
\title{Validate user entry to test based on session token look up in PostgreSQL DB}
\usage{
validate_user_entry_into_test(validate_user_entry_into_test, elts, ...)
}
\arguments{
\item{validate_user_entry_into_test}{Logical or a function which accepts username and ... as arguments and returns a page to present to the user.}

\item{elts}{}
}
\description{
Validate user entry to test based on session token look up in PostgreSQL DB
}
