% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{compile_item_trials}
\alias{compile_item_trials}
\title{Compile item trials}
\usage{
compile_item_trials(
  db_con = NULL,
  current_test_id = NULL,
  session_id = NULL,
  user_id,
  join_item_banks_on = FALSE,
  filter_item_banks = NULL,
  add_trial_scores = FALSE,
  score_to_use = "opti3"
)
}
\arguments{
\item{score_to_use}{}
}
\description{
Compile item trials
}
