% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_append_session.R,
%   R/lambda_append_session_failed.R
\name{db_append_session}
\alias{db_append_session}
\title{Store session information in sessions table}
\usage{
db_append_session(
  db_con,
  experiment_condition_id = NA,
  user_id,
  psychTestR_session_id = NA,
  session_time_started = Sys.time(),
  experiment_id = NA,
  user_info = NA
)

db_append_session(
  db_con,
  experiment_condition_id = NA,
  user_id,
  psychTestR_session_id = NA,
  session_time_started = Sys.time(),
  experiment_id = NA,
  user_info = NA
)
}
\arguments{
\item{user_info}{}
}
\description{
Store session information in sessions table

Store session information in sessions table
}
