% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_append_session.R
\name{db_append_session}
\alias{db_append_session}
\title{Store session information in sessions table}
\usage{
db_append_session(
  db_con,
  condition_id = NA,
  user_id,
  psychTestR_session_id,
  session_time_started = Sys.time(),
  experiment_id = NA
)
}
\arguments{
\item{experiment_id}{}
}
\description{
Store session information in sessions table
}
