% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_add_trial_and_compute_trial_scores.R
\name{db_append_trials}
\alias{db_append_trials}
\title{Append a trial to the trials table}
\usage{
db_append_trials(
  db_con,
  audio_file,
  trial_time_started,
  trial_time_completed,
  instrument,
  attempt,
  item_id,
  display_modality,
  phase,
  rhythmic,
  session_id,
  test_id,
  stimulus_abs_melody,
  stimulus_durations,
  review_items_id = NULL,
  new_items_id = NULL,
  trial_type = c("audio", "midi"),
  trial_paradigm = c("call_and_response", "simultaneous_recall",
    "long_note_call_and_response", "long_note_simultaneous_recall",
    "setup_sing_range_note"),
  additional = NA,
  melody_block_paradigm,
  page_label,
  module = ""
)
}
\arguments{
\item{module}{}
}
\description{
Append a trial to the trials table
}
